+++ scripts/build-windows.sh
@@ -1,3 +1,6 @@
+# Deprecated for now.
+# We should later use it for portable version of Halloy.
+
 #!/bin/bash
 EXE_NAME="halloy.exe"
 TARGET="x86_64-pc-windows-msvc"